refactor: cron scheduler

+++ social-media-app/server/src/config/cronjobScheduler.ts
@@ -1,6 +1,5 @@
 import cron from "node-cron";
 import { scheduledPostSchema } from "../models/scheduledPostModel";
-import { deletePost, savePost } from "../service/postService";
 import postSchema from "../models/postModel";
 
 // Run every 2 minute
@@ -11,9 +10,13 @@ const task = cron.schedule(
     const scheduledPosts = await scheduledPostSchema.find({
       time: { $lte: now.toISOString() },
     });
-
-    await postSchema.insertMany(scheduledPosts);
-    await scheduledPostSchema.deleteMany({ time: { $lte: now.toISOString() } });
+    if(scheduledPosts.length === 0){
+      task.stop();
+    }
+    else{
+      await postSchema.insertMany(scheduledPosts);
+      await scheduledPostSchema.deleteMany({ time: { $lte: now.toISOString() } });
+    }
   },
   {
     scheduled: false,
